Notification test: Push and Email
